aula 32

+++ src/app/heroes/components/heroes/heroes.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from 'src/app/core/components/confirmation-dialog/confirmation-dialog.component';
+import { DialogData } from 'src/app/core/models/dialog-data-model';
 import { Hero } from '../../../core/models/hero.model';
 import { HeroService } from '../../../core/services/hero.service';
 
@@ -14,7 +17,7 @@ export class HeroesComponent implements OnInit {
   heroes: Hero[] = [];
 
 
-  constructor(private heroService: HeroService) {}
+  constructor( private dialog: MatDialog, private heroService: HeroService) {}
 
   ngOnInit(): void {
     this.getHeroes();
@@ -26,10 +29,30 @@ export class HeroesComponent implements OnInit {
   }
 
   delete(hero: Hero): void {
-    this.heroService.delete(hero).subscribe(() => {
-      this.heroes = this.heroes.filter(h => h !== hero)
-    })
+    const dialogData: DialogData = {
+      cancelText: "Cancelar",
+      confirmText: "Confirmar",
+      content: ` Deseja Apagar o Herói ${hero.name}?`,
+    };
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: dialogData,
+      width: '300px',
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result){
+
+        this.heroService.delete(hero).subscribe(() => {
+          // this.heroes = this.heroes.filter(h => h !== hero);
+          this.getHeroes();
+        });
+      }
+    });
+
   }
 
-
+  onSelected(hero: Hero): void {
+    this.delete(hero);
+  }
 }
